+++ truffle/src/com.oracle.truffle.api.dsl.test/src/com/oracle/truffle/api/dsl/test/NodeFieldTest.java
@@ -42,9 +42,12 @@ package com.oracle.truffle.api.dsl.test;
 
 import static com.oracle.truffle.api.dsl.test.TestHelper.createCallTarget;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertSame;
 
 import org.junit.Test;
 
+import com.oracle.truffle.api.dsl.GenerateUncached;
 import com.oracle.truffle.api.dsl.NodeField;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.dsl.test.NodeFieldTestFactory.IntFieldNoGetterTestNodeFactory;
@@ -54,7 +57,12 @@ import com.oracle.truffle.api.dsl.test.NodeFieldTestFactory.ObjectContainerNodeF
 import com.oracle.truffle.api.dsl.test.NodeFieldTestFactory.RewriteTestNodeFactory;
 import com.oracle.truffle.api.dsl.test.NodeFieldTestFactory.StringFieldTestNodeFactory;
 import com.oracle.truffle.api.dsl.test.NodeFieldTestFactory.TestContainerFactory;
+import com.oracle.truffle.api.dsl.test.NodeFieldTestFactory.UncachedNodeIntFieldRefNodeGen;
+import com.oracle.truffle.api.dsl.test.NodeFieldTestFactory.UncachedNodeIntFieldTestNodeGen;
+import com.oracle.truffle.api.dsl.test.NodeFieldTestFactory.UncachedNodeObjectFieldTestNodeGen;
 import com.oracle.truffle.api.dsl.test.TypeSystemTest.ValueNode;
+import com.oracle.truffle.api.nodes.Node;
+import com.oracle.truffle.api.source.SourceSection;
 
 public class NodeFieldTest {
 
@@ -185,6 +193,77 @@ public class NodeFieldTest {
             return getObject();
         }
 
+    }
+
+    @Test
+    public void testUncachedNodeIntFieldTest() {
+        assertEquals(42, UncachedNodeIntFieldTestNodeGen.create(42).execute());
+        assertEquals(0, UncachedNodeIntFieldTestNodeGen.getUncached().execute());
+    }
+
+    @GenerateUncached
+    @NodeField(name = "field", type = int.class)
+    public abstract static class UncachedNodeIntFieldTest extends Node {
+
+        abstract Object execute();
+
+        protected abstract int getField();
+
+        @Specialization
+        Object s0() {
+            return getField();
+        }
+
+        @Override
+        public SourceSection getSourceSection() {
+            // uses getSourceLength
+            return null;
+        }
+    }
+
+    @Test
+    public void testUncachedNodeObjectFieldTest() {
+        Object instance = new Object();
+        assertSame(instance, UncachedNodeObjectFieldTestNodeGen.create(instance).execute());
+        assertNull(UncachedNodeObjectFieldTestNodeGen.getUncached().execute());
+    }
+
+    @GenerateUncached
+    @NodeField(name = "field", type = Object.class)
+    public abstract static class UncachedNodeObjectFieldTest extends Node {
+
+        abstract Object execute();
+
+        protected abstract Object getField();
+
+        @Specialization
+        Object s0() {
+            return getField();
+        }
+
+        @Override
+        public SourceSection getSourceSection() {
+            // uses getSourceLength
+            return null;
+        }
+    }
+
+    @Test
+    public void testUncachedNodeIntFieldRef() {
+        assertEquals(42, UncachedNodeIntFieldRefNodeGen.create(42).execute());
+        assertEquals(0, UncachedNodeIntFieldRefNodeGen.getUncached().execute());
+    }
+
+    @GenerateUncached
+    @NodeField(name = "foo", type = int.class)
+    public abstract static class UncachedNodeIntFieldRef extends Node {
+
+        abstract Object execute();
+
+        @Specialization
+        static Object s0(int foo) {
+            return foo;
+        }
     }
 
 }